Started work on SPCTRM
- IDEALLY will be an interactive spectrum (as the name implies hoho see
what I did there) for gender, sexuality, and romance (as inspired by
the asexual community), hopefully in a very tactful manner
- Spent, like, a really long time doing a rainbow gradient which is sad
because it's not that hard
- Also made TextObj class because there are just so many gosh darn
things to keep track of, mane.
- Also realized that having "standard.js" is a horrible idea because
I'm not nearly well-organized enough—or experience enough—to maintain
anything close to library...so, like, probably just never going to
attempt that again. Copy-paste ftw!
- ... same goes for having a cordoned off canvas projects into their
own little folder...moved SPCTRM out into the open, the wild!
- M E D I A